[Feat] #41 - 마이페이지 레이아웃 구현 완료

+++ app/src/main/java/com/example/zupzup/data/datasource/reservation/ReservationDataSource.kt
@@ -4,5 +4,5 @@ import com.example.zupzup.data.dto.Reservation
 
 interface ReservationDataSource {
 
-    suspend fun createReservation(reservation: Reservation): Result<Int>
+    suspend fun createReservation(reservation: Reservation): Result<Long>
 }
